making progress with brand endpoints

+++ src/categories.js
@@ -2,6 +2,7 @@ const express = require('express');
 const router = express.Router();
 
 const Category = require('./models/category');
+const Brand = require('./models/brand');
 
 //  get to /category/:name returns the category with name = :name
 router.get("/:name", (req, res, next) => {
@@ -38,6 +39,7 @@ router.put("", (req, res, next) => {
 });
 
 // delete to /category/:name removes the category with name = :name
+// and all the brands of such a category
 router.delete("/:name", (req, res, next) => {
     remove(req.params.name)
     .then(
@@ -58,7 +60,6 @@ function search(name) {
             if (!category) {
                 return `No category with name ${name}`;
             }
-            console.log(category);
             return category;
         },
         err => { throw new Error(err) }
@@ -73,10 +74,6 @@ function add(name, description) {
                 return `Category with name ${name} already exists`;
             }
             return Category.create({name:name, description: description })
-            .then(
-                category => { return category },
-                err => { throw new Error(err) }
-            )
         },  
         err => { throw new Error(err) }  
     )
@@ -91,19 +88,25 @@ function edit(name, description) {
             }
             category.description = description;
             return category.save()
-            .then(
-                category => { return category },
-                err => { throw new Error(err) }
-            );
         },
         err => { throw new Error(err) }
     )
 }
 
 function remove(name) {
-    return Category.deleteOne({name:name})
+    return Category.findOne({name:name})
     .then(
-        () => { return `Category ${name} deleted` },
-        err => { throw new Error(err) }
+        category => {
+            if (!category) {
+                return `No category with name ${name}`;
+            }
+            Brand.deleteMany({category: category._id}).exec();
+            return Category.deleteOne({name:name})
+            .then(
+                () => { return `Category ${name} deleted` },
+                err => { throw new Error(err) }
+            )
+        }
     )
+    
 }
